% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{beta.hat.func}
\alias{beta.hat.func}
\title{Function of Beta.hat}
\usage{
beta.hat.func(X, Z, Y, setting)
}
\arguments{
\item{X}{distorted covariates}

\item{Z}{observed covariates}

\item{Y}{distorted response}

\item{setting}{1-4 represent model (5.1-5.4)}
}
\value{
estimates of beta and residual
}
\description{
Obtain the Least Square estimate
}
\examples{
beta_e.hat = beta.hat.func(X.tilde, Z, Y.tilde, 1)
}
