% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{bootstrap.res}
\alias{bootstrap.res}
\title{Bootstrap procedure}
\usage{
bootstrap.res(X.hat, Z, Y.hat, e.hat, setting, rho)
}
\arguments{
\item{rho}{the number of Bootstrap replications}
}
\value{
the Bootstrap residuals, a n*rho matrix
}
\description{
Bootstrap procedure
}
\examples{
e.star = bootstrap.res(X.hat, Z, Y.hat, e.hat, 1, rho = 1000)
}
