% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{cali.func}
\alias{cali.func}
\title{Calibrating function}
\usage{
cali.func(x, u, kern)
}
\arguments{
\item{x}{a n*1 vector}

\item{u}{a n*1 vector}

\item{kern}{kernel parameters}
}
\value{
a n*1 calibrated vector
}
\description{
Calibrating function
}
\examples{
Y.hat = cali.func(Y.tilde, U, kerns)
}
