% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{DGP}
\alias{DGP}
\title{Data generate process}
\usage{
DGP(n, C, setting)
}
\arguments{
\item{n}{sample size}

\item{C}{the difference bewteen H0 and H1}

\item{setting}{1-4 represent model (5.1-5.4)}
}
\value{
a list contains X.tilde, Y.tilde, Z, and U.
}
\description{
Data generate process
}
\examples{
data = DGP(n = 10, C = 0, setting = 1)
}
