% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{Tcvm}
\alias{Tcvm}
\title{Calculate Tcvm}
\usage{
Tcvm(X, Z, e)
}
\arguments{
\item{X}{X.hat}

\item{e}{e.hat}
}
\value{
Tcvm test statistics(2.4)
}
\description{
Calculate Tcvm
}
\examples{
Tcvm(X.hat, Z, e.hat)
}
