% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{Tks.rand}
\alias{Tks.rand}
\title{Calculate Tks.rand}
\usage{
Tks.rand(V, theta, e)
}
\arguments{
\item{V}{cbind(X.hat, Z) a n*(p+q) matrix}

\item{e}{e.hat}
}
\value{
Tks test statistics(3.1)
}
\description{
Calculate Tks.rand
}
\examples{
Tks.rand(V, theta.normal, e.hat)
}
