% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tks_rand_rcpp}
\alias{Tks_rand_rcpp}
\title{Calculate Tks.rand (rcpp version)}
\usage{
Tks_rand_rcpp(V, theta, e)
}
\arguments{
\item{V}{cbind(X.hat, Z) a n*(p+q) matrix}

\item{e}{e.hat}
}
\value{
Tks test statistics(3.1)
}
\description{
Calculate Tks.rand (rcpp version)
}
\examples{
Tks_rand_rcpp(V, theta.normal, e.hat)
}
